backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "alembic==1.16.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "build==1.3.0",
    "cachetools==6.2.4",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "chromadb==1.4.0",
    "click==8.2.1",
    "cloudinary==1.44.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "cryptography==45.0.4",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "durationpy==0.10",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.13",
    "fastapi-cli==0.0.7",
    "filelock==3.20.1",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "frozenlist==1.7.0",
    "fsspec==2025.12.0",
    "geoip2==5.1.0",
    "google-auth==2.45.0",
    "google-genai==1.56.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.3",
    "grpcio==1.76.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==34.1.0",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-google-genai==4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters==1.1.0",
    "langsmith==0.5.1",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.2",
    "maxminddb==2.8.2",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "numpy!=2.4.0",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "posthog==5.4.0",
    "propcache==0.3.2",
    "protobuf==6.33.2",
    "psycopg2==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.3",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-extra-types==2.10.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "resend==2.11.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "sentence-transformers>=5.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.1",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typer-slim==0.21.0",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "ujson==5.10.0",
    "urllib3==2.3.0",
    "uuid-utils==0.12.0",
    "uvicorn==0.34.3",
    "watchfiles==1.1.0",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
